requirements.txt
python-dotenv==1.0.0
praw==7.7.1
finnhub-python==2.4.18
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
pandas==2.2.0
numpy==1.26.4
pytest==8.0.0
requests==2.31.0
python-dateutil==2.8.2 